reexec: drop Pdeathsig
it can lead to a race condition where the process is killed even if
the parent process is still running.

It can happen as the parent in the prctl(PR_SET_PDEATHSIG) case is
considered the *thread* that created the process.  In the case of the
Go runtime the thread could be freed by the runtime at any time
causing the re-execed process to receive the SIGTERM.

Signed-off-by: Giuseppe Scrivano <gscrivan@redhat.com>