pkg/mount.ensureMountedAs: simplify
1. There is no need to specify rw argument -- bind mounts are
read-write by default.

2. Use ForceMount() instead of Mount() to avoid double parsing
of /proc/self/mountinfo -- we've just checked this is not a
mount point.

3. There is no point in parsing /proc/self/mountinfo after performing
a mount, especially if we don't check whether the fs is mounted or
not -- the only outcome from it could be an error from our mountinfo
parser, which makes no sense in this context.

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>