Downgrade logrus to v1.4.2
The upgrade of logrus forces to update the dependency in go module
consumers as well.

The new locking in logrus 1.5.0 introduces regressions in conjunction to
deadlocks when using custom hooks, which is the case in CRI-O for
example.

Let's downgrade logrus until the [referencing issue is
fixed](https://github.com/sirupsen/logrus/issues/1122).

Signed-off-by: Sascha Grunert <sgrunert@suse.com>