overlay: recreate workdir on UpdateLayerIDMap
make sure the work dir is also recreated with the diff dir.

When using index=on, overlay records a reference to the upperdir inode
into work/index, and raises an error if on a successive mount there is
a mismatch between what is stored in the work/index xattr and the
upper dir inode.

Closes: https://github.com/containers/libpod/issues/5816

Signed-off-by: Giuseppe Scrivano <gscrivan@redhat.com>