Specify default paths to policy.json file
Currently if a rootless user runs podman in his homedir, and does
not have /etc/containers/policy.json on the system and does not
have ~/config/policy.json in his homedir, the command fails with
an error.

$ podman pull fedora
Error: error pulling image "fedora": unable to pull fedora: open /etc/containers/policy.json: no such file or directory

If the user has no root rights on the system, he can not create the file.
However the system would work fine if he created the file in his homedir.

With this change, we will force the default to be the file in his homedir if
BOTH files do not exist.  Now the error message for a rootless container would
be.

$ podman pull fedora
Error: unable to pull fedora: open /home/dwalsh/.config/containers/policy.json: no such file or directory

And the user has an idea how to fix this situation.

Signed-off-by: Daniel J Walsh <dwalsh@redhat.com>