Ignore empty strings from libpod.conf
If the user has a libpod.conf file with an option of "", then
we will ignore it and use the value from containers.conf.

Also the Set variables are only to be set, if a user specified an
override command, not if they were set in the libpod.conf file.

Signed-off-by: Daniel J Walsh <dwalsh@redhat.com>