oci: don't overwrite tags pointing at the same manifest
In 78fab78893f7 ("Add manifest list support"), the code was changed from a
single step (match on the name annotation) to a two step process: 1. untag
the existing name, and 2. match on the manifest's digest and change the tag
of this match to the new tag.

However, this means that if there were two tags:

    "foo": sha256:c0ffee55
    "bar": sha256:deadbeef

And we update "bar" to point to c0ffee55, we'd end up doing:

1. remove any "bar" tags:

    "foo": sha256:c0ffee55
    "": sha256:deadbeef

2. re-tag the first c0ffee55 thing:

    "bar": sha256:c0ffee55
    "": sha256:deadbeef

Thus losing the "foo" tag entirely.

Instead, let's make sure that we only re-tag something with the empty
string, so we don't delete existing tags.

Signed-off-by: Tycho Andersen <tycho@tycho.ws>