Use a loop to go through repos in quay script
While trying to run down an issue with cron, I releasized that the
only differences between the steps for the stable, testing and upstream
was the name of the repo.  So I put the builds of those three
flavors under a for loop and shrunk the script down quite a bit.

Signed-off-by: TomSweeneyRedHat <tsweeney@redhat.com>