store: support mapped layers deletion
currently it is not possible to delete a mapped layer as it is always
referenced by the image.

Change it allow deleting the mapping if it is not used by any
container or other layers.

It will be useful in CRI-O as a lot of thrown away containers will be
created and the mapped images would keep growing without a way of
cleaning them up.

Signed-off-by: Giuseppe Scrivano <gscrivan@redhat.com>