Tighten permissions on created directory
The mount points of containers are being marked as 755 where
a lot of images now come with the default for "/", as 555.

The tighter controls on / means that a root process can not
modify the / directory without DAC_OVERRIDE.  And some consider
this better security.  Bottom line is tools like `rpm -qVf /`
are reporting failures on permissions on / in lots of RPM
based images.

Signed-off-by: Daniel J Walsh <dwalsh@redhat.com>