Cleanup code to explain what is going on here.
I got fooled by this code, and attempted a PR to clean it up, that
was a mistake.  I assumed that the call should fail if the Container
call failed. But the store.Mount(id, mountlabel) handles container
mounts as well as layer mounts. This means the ID could be a container
or a different layer. When the attempt to read the container fails, we
continue assuming that the id is a layer and attempt to mount it as a layer.
If the id is a container then we grap the toplayer of the container and use
it is as an id.

This PR does not change the functionality of the code, but makes it easier
for future programmers to understand the functionality.

Signed-off-by: Daniel J Walsh <dwalsh@redhat.com>