Move the allocation of an empty alias map to editShortNameAlias
This is the only function that actually cares about the difference
between nil and an empty non-nil map, so do it here. It's easier to
maintain when the check has a clear motivaction (and it avoids an
unnecessary allocation on read-only paths, though that alone is
a pretty bad reason to move the code).

Should not change behavior.

Signed-off-by: Miloslav Trmač <mitr@redhat.com>