Behavior change: Move unqualifiedSearchRegistriesOrigin to parsedConfig
In addition to the simple move, get closer to the load+merge pattern:
loadConfigFile now sets path, and loadConfig updates it based on the
merge decision.

This changes behavior if no config files set unqualified-search-registries:
before, the empty value would be attributed to first loaded file (whether .conf
or .d), now the origin is always "".

Signed-off-by: Miloslav Trmač <mitr@redhat.com>