overlay: set permissions on "diff" correctly
When creating a new layer that has a parent, assign the new layer's
"diff" directory the same permissions as the parent layer's "diff"
directory, rather than the permissions of the parent of its parent
layer's "diff" directory.

Before the recent round of changes, that directory would be initialized
with 0700 permissions, so layers created by new code on top of layers
created by the older version would be given 0700 permissions, and were
not usable by UIDs other than 0.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>