Implement secrets pkg: backend and filedriver
This is the implementation of the backend of secrets. pkg/secrets takes a secret name and data and does these operations on that secret data:  store, delete, lookup, and list, using a secretsmanager.

The first driver implemented here is a filedriver - where the data is stored unencrypted on disk in a file.

The secrets package can be easily expanded to use more drivers as the package implements an interface to accept different drivers

Signed-off-by: Ashley Cui <acui@redhat.com>