Skip pulling layers reusable from additional layer store
This commit adds support for "Additional Layer Store".

Pull is one of the time-consuming steps in the container lifecycle. Additional
Layer Store enables runtimes (e.g. Podman, CRI-O, etc) to startup containers
using layers stored in a specified directory, instead of pulling them from the
registry.

One of the expected use cases of this feature is "lazy pulling". This enables
the runtimes to startup containers without waiting for the entire image contents
to be locally available, but necessary chunks of the contents are fetched
on-demand (lazily).

There are several image formats and filesystems to enable lazy pulling in the
community. They includes stargz/eStargz, zstd:chunked, CVMFS, etc. Additional
Layer Store makes it easy to integrate with these filesystems for performing
lazy pulling.

Signed-off-by: Kohei Tokunaga <ktokunaga.mail@gmail.com>