pkg/lockfile: fix a race and an incorrect unit test
Fix a race where the state lock in a Unixy lockfile structure wasn't
being used to protect access to all of the structure's state fields.

Fix usage of sync/atomic in the corresponding unit tests to not mix
atomic adds and reads with non-atomic reads, something which the race
detector complained about.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>