drivers/zfs: default the base layer to 0555
When creating a new zfs base layer, set its permissions to 0555 instead
of leaving it at the default of apparently 0755, bringing it in line
with overlay.  Since we need to mount it to do that, use the opportunity
to set the ownership, which we previously did (and re-did) every time we
mounted a volume after it was created.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>